+++ Planarian.Web/src/Modules/Import/Components/ImportEntrances.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { Card, Result, Button, Modal, message, Spin } from "antd";
+import { Card, Result, Button, Modal, message } from "antd";
 import {
   DeliveredProcedureOutlined,
   CheckCircleOutlined,
@@ -10,7 +10,6 @@ import Papa from "papaparse";
 // Importing components and services
 import { UploadComponent } from "../../Files/Components/UploadComponent";
 import { CSVDisplay } from "../../Files/Components/CsvDisplayComponent";
-import { CaveService } from "../../Caves/Service/CaveService";
 
 // Importing models
 import {
@@ -26,6 +25,7 @@ import { PlanarianButton } from "../../../Shared/Components/Buttons/PlanarianBut
 import { NotificationComponent } from "./NotificationComponent";
 import { FullScreenModal } from "../../Files/Components/FileListItemComponent";
 import { Link } from "react-router-dom";
+import { AccountService } from "../../Account/Services/AccountService";
 
 interface ImportEntrancesComponentProps {
   onUploaded: () => void;
@@ -67,7 +67,7 @@ const ImportEntrancesComponent: React.FC<ImportEntrancesComponentProps> = ({
     setIsLoading(true);
     setIsProcessing(true); // Set processing state to true when starting to process
     try {
-      await CaveService.ImportEntrancesProcess(uploadResult.id);
+      await AccountService.ImportEntrancesProcess(uploadResult.id);
       setIsProcessed(true);
     } catch (e) {
       const error = e as ImportApiErrorResponse<EntranceCsvModel>;
@@ -104,7 +104,7 @@ const ImportEntrancesComponent: React.FC<ImportEntrancesComponentProps> = ({
           style={{ display: "flex" }}
           uploadFunction={async (params): Promise<FileVm> => {
             try {
-              const result = await CaveService.ImportEntrancesFile(
+              const result = await AccountService.ImportEntrancesFile(
                 params.file,
                 params.uid,
                 params.onProgress
